Add access control to feed crud

+++ src/Controller/FeedCrudController.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 namespace App\Controller;
 
 use App\Entity\Feed;
+use App\Entity\RssUser;
 use App\Feed\Processor as FeedProcessor;
 use App\Form\FeedType;
 use App\Repository\FeedRepository;
+use App\Services\Permissions;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -14,17 +16,21 @@ use Symfony\Component\Routing\Annotation\Route;
 
 /**
  * @Route("/admin/feed")
+ *
+ * @method RssUser getUser()
  */
 class FeedCrudController extends AbstractController
 {
-    /**
-     * @var FeedProcessor
-     */
+    /** @var FeedProcessor */
     private $feedProcessor;
 
-    public function __construct(FeedProcessor $feedProcessor)
+    /** @var Permissions */
+    private $permissions;
+
+    public function __construct(FeedProcessor $feedProcessor, Permissions $permissions)
     {
         $this->feedProcessor = $feedProcessor;
+        $this->permissions   = $permissions;
     }
 
     /**
@@ -87,8 +93,8 @@ class FeedCrudController extends AbstractController
         $form = $this->createForm(FeedType::class, $feed);
         $form->handleRequest($request);
 
-        if ($feed->getRssUser() !== $this->getUser()) {
-            throw $this->createAccessDeniedException();
+        if ($this->permissions->canEditContentFromUser($this->getUser(), $feed->getRssUser()) === false) {
+            throw $this->createAccessDeniedException('Cannot edit a feed that is not yours.');
         }
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -109,6 +115,10 @@ class FeedCrudController extends AbstractController
     public function delete(Request $request, Feed $feed): Response
     {
         if ($this->isCsrfTokenValid('delete' . $feed->getId(), $request->request->get('_token'))) {
+            if ($this->permissions->canEditContentFromUser($this->getUser(), $feed->getRssUser()) === false) {
+                throw $this->createAccessDeniedException('Cannot delete a feed that is not yours.');
+            }
+
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($feed);
             $entityManager->flush();
